remove stripMargin move model generation to src/test add basic test
